Added cmd-line arguments
Added cmd-line arguments to .ps1 file, so it is possible to change the resolution of the primary monitor only. Used when virtual screens are deactivated and only physical screen is present again.
With this commit, the external program for changing the resolution of the screen through batch-script is now unnecessary, and everything is handled via the .cmd file called from the batch-scripts with arguments for the .ps1 file.